Asset Syncer: Fetch files for postgres
