+++ cmd/asset-syncer/utils.go
@@ -19,6 +19,7 @@ package main
 import (
 	"archive/tar"
 	"bytes"
+	"compress/gzip"
 	"crypto/sha256"
 	"crypto/tls"
 	"crypto/x509"
@@ -31,8 +32,10 @@ import (
 	"os"
 	"path"
 	"strings"
+	"sync"
 	"time"
 
+	"github.com/disintegration/imaging"
 	"github.com/ghodss/yaml"
 	"github.com/jinzhu/copier"
 	"github.com/kubeapps/common/datastore"
@@ -77,6 +80,9 @@ type assetManager interface {
 	UpdateLastCheck(repoName, checksum string, now time.Time) error
 	Init() error
 	Close() error
+	updateIcon(data []byte, contentType, ID string) error
+	filesExist(chartFilesID, digest string) bool
+	insertFiles(chartFilesID string, files chartFiles) error
 }
 
 func newManager(databaseType string, config datastore.Config) (assetManager, error) {
@@ -256,4 +262,193 @@ func initNetClient(additionalCA string) (*http.Client, error) {
 			Proxy: http.ProxyFromEnvironment,
 		},
 	}, nil
+}
+
+type fileImporter struct {
+	manager assetManager
+}
+
+func (f *fileImporter) fetchFiles(charts []chart) {
+	// Process 10 charts at a time
+	numWorkers := 10
+	iconJobs := make(chan chart, numWorkers)
+	chartFilesJobs := make(chan importChartFilesJob, numWorkers)
+	var wg sync.WaitGroup
+
+	log.Debugf("starting %d workers", numWorkers)
+	for i := 0; i < numWorkers; i++ {
+		wg.Add(1)
+		go f.importWorker(&wg, iconJobs, chartFilesJobs)
+	}
+
+	// Enqueue jobs to process chart icons
+	for _, c := range charts {
+		iconJobs <- c
+	}
+	// Close the iconJobs channel to signal the worker pools to move on to the
+	// chart files jobs
+	close(iconJobs)
+
+	// Iterate through the list of charts and enqueue the latest chart version to
+	// be processed. Append the rest of the chart versions to a list to be
+	// enqueued later
+	var toEnqueue []importChartFilesJob
+	for _, c := range charts {
+		chartFilesJobs <- importChartFilesJob{c.Name, c.Repo, c.ChartVersions[0]}
+		for _, cv := range c.ChartVersions[1:] {
+			toEnqueue = append(toEnqueue, importChartFilesJob{c.Name, c.Repo, cv})
+		}
+	}
+
+	// Enqueue all the remaining chart versions
+	for _, cfj := range toEnqueue {
+		chartFilesJobs <- cfj
+	}
+	// Close the chartFilesJobs channel to signal the worker pools that there are
+	// no more jobs to process
+	close(chartFilesJobs)
+
+	// Wait for the worker pools to finish processing
+	wg.Wait()
+}
+
+func (f *fileImporter) importWorker(wg *sync.WaitGroup, icons <-chan chart, chartFiles <-chan importChartFilesJob) {
+	defer wg.Done()
+	for c := range icons {
+		log.WithFields(log.Fields{"name": c.Name}).Debug("importing icon")
+		if err := f.fetchAndImportIcon(c); err != nil {
+			log.WithFields(log.Fields{"name": c.Name}).WithError(err).Error("failed to import icon")
+		}
+	}
+	for j := range chartFiles {
+		log.WithFields(log.Fields{"name": j.Name, "version": j.ChartVersion.Version}).Debug("importing readme and values")
+		if err := f.fetchAndImportFiles(j.Name, j.Repo, j.ChartVersion); err != nil {
+			log.WithFields(log.Fields{"name": j.Name, "version": j.ChartVersion.Version}).WithError(err).Error("failed to import files")
+		}
+	}
+}
+
+func (f *fileImporter) fetchAndImportIcon(c chart) error {
+	if c.Icon == "" {
+		log.WithFields(log.Fields{"name": c.Name}).Info("icon not found")
+		return nil
+	}
+
+	req, err := http.NewRequest("GET", c.Icon, nil)
+	if err != nil {
+		return err
+	}
+	req.Header.Set("User-Agent", userAgent())
+	if len(c.Repo.AuthorizationHeader) > 0 {
+		req.Header.Set("Authorization", c.Repo.AuthorizationHeader)
+	}
+
+	res, err := netClient.Do(req)
+	if res != nil {
+		defer res.Body.Close()
+	}
+	if err != nil {
+		return err
+	}
+
+	if res.StatusCode != http.StatusOK {
+		return fmt.Errorf("%d %s", res.StatusCode, c.Icon)
+	}
+
+	b := []byte{}
+	contentType := ""
+	if strings.Contains(res.Header.Get("Content-Type"), "image/svg") {
+		// if the icon is a SVG file simply read it
+		b, err = ioutil.ReadAll(res.Body)
+		if err != nil {
+			return err
+		}
+		contentType = res.Header.Get("Content-Type")
+	} else {
+		// if the icon is in any other format try to convert it to PNG
+		orig, err := imaging.Decode(res.Body)
+		if err != nil {
+			log.WithFields(log.Fields{"name": c.Name}).WithError(err).Error("failed to decode icon")
+			return err
+		}
+
+		// TODO: make this configurable?
+		icon := imaging.Fit(orig, 160, 160, imaging.Lanczos)
+
+		var buf bytes.Buffer
+		imaging.Encode(&buf, icon, imaging.PNG)
+		b = buf.Bytes()
+		contentType = "image/png"
+	}
+
+	return f.manager.updateIcon(b, contentType, c.ID)
+}
+
+func (f *fileImporter) fetchAndImportFiles(name string, r *repo, cv chartVersion) error {
+	chartFilesID := fmt.Sprintf("%s/%s-%s", r.Name, name, cv.Version)
+
+	// Check if we already have indexed files for this chart version and digest
+	if f.manager.filesExist(chartFilesID, cv.Digest) {
+		log.WithFields(log.Fields{"name": name, "version": cv.Version}).Debug("skipping existing files")
+		return nil
+	}
+	log.WithFields(log.Fields{"name": name, "version": cv.Version}).Debug("fetching files")
+
+	url := chartTarballURL(r, cv)
+	req, err := http.NewRequest("GET", url, nil)
+	if err != nil {
+		return err
+	}
+	req.Header.Set("User-Agent", userAgent())
+	if len(r.AuthorizationHeader) > 0 {
+		req.Header.Set("Authorization", r.AuthorizationHeader)
+	}
+
+	res, err := netClient.Do(req)
+	if err != nil {
+		return err
+	}
+	defer res.Body.Close()
+
+	// We read the whole chart into memory, this should be okay since the chart
+	// tarball needs to be small enough to fit into a GRPC call (Tiller
+	// requirement)
+	gzf, err := gzip.NewReader(res.Body)
+	if err != nil {
+		return err
+	}
+	defer gzf.Close()
+
+	tarf := tar.NewReader(gzf)
+
+	readmeFileName := name + "/README.md"
+	valuesFileName := name + "/values.yaml"
+	schemaFileName := name + "/values.schema.json"
+	filenames := []string{valuesFileName, readmeFileName, schemaFileName}
+
+	files, err := extractFilesFromTarball(filenames, tarf)
+	if err != nil {
+		return err
+	}
+
+	chartFiles := chartFiles{ID: chartFilesID, Repo: r, Digest: cv.Digest}
+	if v, ok := files[readmeFileName]; ok {
+		chartFiles.Readme = v
+	} else {
+		log.WithFields(log.Fields{"name": name, "version": cv.Version}).Info("README.md not found")
+	}
+	if v, ok := files[valuesFileName]; ok {
+		chartFiles.Values = v
+	} else {
+		log.WithFields(log.Fields{"name": name, "version": cv.Version}).Info("values.yaml not found")
+	}
+	if v, ok := files[schemaFileName]; ok {
+		chartFiles.Schema = v
+	} else {
+		log.WithFields(log.Fields{"name": name, "version": cv.Version}).Info("values.schema.json not found")
+	}
+
+	// inserts the chart files if not already indexed, or updates the existing
+	// entry if digest has changed
+	return f.manager.insertFiles(chartFilesID, chartFiles)
 }
